requirements.txt
Flask==2.3.2
numpy==1.24.3
opencv_python==4.7.0.72
Pillow==9.5.0